Partial work towards supporting multiple simultaneous series in tiling job
